Create Django project file structure and add .gitignore
This commit creates a Django project called 'Delos' and sets up the basic file structure with .gitignore file. This includes settings, asgi, and wsgi configuration files. A requirements.txt file listing project dependencies and a utility script for system information are also added. Modifications in the README file detail the default packages used in the project.