Implement multi-tenant architecture and create authentication and todo apps
This commit configures the project for a multi-tenant architecture using django-tenants, creating a common 'customers' app to manage tenant information. It also adds the 'authentication' app with a custom user model and manager, and the 'app.todo' as a tenant-specific application. Multiple migrations are included for the newly introduced models. Additionally, Django project structure and settings file have been modified appropriately for the new setup.