Add tenant and domain models to admin, update settings
Added Tenant and Domain models to the admin interface for better management of multi-tenancy. Also, updated the Django settings file, including the creation of the PUBLIC_SCHEMA_URLCONF variable, to accommodate the new structure of the application in line with django_tenants specifications. These changes were necessary to enable a robust multi-tenant architecture in the Django project.