Add "Country" field to Tenant model and update related files
A new field "Country" has been added to the Tenant model in customers/models.py. This required a new migration file to be generated and an update of 'list_display' in customers/admin.py. Also, several new .idea IDE configuration files were added. In Delos/settings.py the `DEBUG` constant value was set to `True` for development and the hard-coded value for `DATA_UPLOAD_MAX_MEMORY_SIZE` was set to 50Mb.