Add initial files for webcore app and update settings
This commit introduces the initial files for the webcore application, including views, tests, models, and URLs. The settings in Delos/settings.py file were also adjusted: the inclusion of 'webcore' in the installed apps and updating the ALLOWED_HOSTS to include 'localhost'. The admin route in Delos/urls.py was removed and added to Delos/urls_public.py along with the new webcore URLs.