Refactor webcore templates and update settings
This commit refactors the index.html template to extend from a base template, with dynamic title, script, and content blocks. It also introduces a new navbar.html template. Moreover, the setting for BASE_DIR is updated in settings.py, and 'ezeflow.localhost' is added to ALLOWED_HOSTS. The integration of bootstrap_css and bootstrap_javascript into the base.html is also part of this update.