Add Teams model and adjust related fields
Added a new `Teams` model to the `customers` application, including a migration. Altered the `User` model, by adding `birth_date` field and updating `team` field related to `Teams` model. `Tenant` model was also updated with removal of `contact` and `teams` fields and addition of `is_active` field. The changes required a range of migrations to be adjusted.